[Rename] #4 - 파일 변경 및 변수명 변경 적용
